Bug 515373 - Prepare target platform for Papyrus Compare
With this change, we switch to a custom build of EMF Compare that also
provides the test framework and includes a couple of other patches that
are relevant to Papyrus Compare. Once these have been merged into the
main stream of EMF Compare, we can switch back to the official Oxygen
stream of EMF Compare. However, this may take a while.

Change-Id: Ib63cfa055a8e99dfc55fcecaac136a7fde7d5e54
Signed-off-by: Philip Langer <planger@eclipsesource.com>