Improve the item to be selected computation for large trees
For a null itemValue, no processing is needed, so avoid visiting the
whole tree in this case. Sometimes the item value isn't in the tree,
resulting in the whole tree being visited. Limit the amount of time
spent visiting, and when that time expires, instead selected the closest
container that is in the tree.

Change-Id: Ib6aca3d31c5ae43b46ca22b295867829c7c8bd7c
Signed-off-by: Philip Langer <planger@eclipsesource.com>