Bug 528272: [Tests] Profile migration tests are disabled
Create a private profile containing the absolute minimum required
by these tests, mirroring the SysML profile.  Use this profile to
re-create all of the test cases.

Other changes required for correct operation of the migration
include a 'read-only handler' for the test context ensuring that
the profile-migration hook does not see test resources as being
unwritable and therefore skipping migration of their stereotypes.

https://bugs.eclipse.org/bugs/show_bug.cgi?id=528272

Change-Id: I43e51fda61136eb0802fb0322b665c4611b5851c