Bug 529253 - [Compare] Stereotype applications left behind by sub-unit refactoring merge
Add a post-processor that analyzes ResourceAttachmentChanges
to detect incoming changes that are actually refactorings
(rename/move) of resources.  Re-write the MatchResources
accordingly to record the matching of resources on the two
sides that have different URIs, and use this in the merge
phase to ensure that the merge result reflects the new
resource URIs.

This includes JUnit tests for that scenario specifically
and also others in the scope of bug 529217.

Change-Id: Iec8a2a9c8169f05000740383691978d9100b91e6