Bug 529413 - Add model inclusion tester for Papyrus properties files
Previously EMF Compare already included properties-files when computing
the logical model (EMFModelProvider). With the fix for bug 529413,
however, EMF Compare stops doing that and instead provides an extension
point to allow third-party plug-ins to add their file extensions that
they want to be handled by the EMFModelProvider.

This change introduces an extension for this extension point, so that
di-files, sash-files, and properties-files are included. Note that UML
and notation files don't need to be included here, since they are
included by EMF Compare's plug-ins already anyway.

Because neither the EMF Compare project nor the Logical Modeling
fork of it publishes the org.eclipse.compare.uml2.ide bundle in any
feature, the Papyrus Compare test bundles that execute merge/rebase
tests in a git context have to provide their own extension for UML
resources on the oee.compare.ide.modelInclusionTester point.

Change-Id: I90d5410ca7ed31d4aa098b1bbdc5b67a0c638932
Signed-off-by: Philip Langer <planger@eclipsesource.com>
Also-by: Christian W. Damus <give.a.damus@gmail.com>