Bug 529253 - [Compare] Stereotype applications left behind by sub-unit refactoring merge
Rework the support for resource refactoring using an adapter
that marks a ResourceAttachmentChange as representing a
resource refactoring change, making its special semantics
explicit.  A merger for that specific diff takes care of the actual
refactoring.  An item-provider decorator customizes its
presentation in the UI to clearly show that it is a resource
refactoring and not just a root element moved to another
resource.

Change-Id: I0e75c58b0926de240507ddce25c773c9b24eb1d7